refactor: move logtable out of scheduletab for modularity
