+++ src/vorta/assets/UI/scheduletab.ui
@@ -581,66 +581,11 @@
       <attribute name="label">
        <string>Log</string>
       </attribute>
-      <layout class="QGridLayout" name="gridLayout_3">
-       <item row="0" column="0">
-        <widget class="QTableWidget" name="logTableWidget">
-         <property name="font">
-          <font>
-           <pointsize>11</pointsize>
-           <bold>false</bold>
-          </font>
-         </property>
-         <property name="showGrid">
-          <bool>false</bool>
-         </property>
-         <property name="sortingEnabled">
-          <bool>true</bool>
-         </property>
-         <attribute name="verticalHeaderVisible">
-          <bool>false</bool>
-         </attribute>
-         <column>
-          <property name="text">
-           <string>Time</string>
-          </property>
-         </column>
-         <column>
-          <property name="text">
-           <string>Category</string>
-          </property>
-         </column>
-         <column>
-          <property name="text">
-           <string>Subcommand</string>
-          </property>
-         </column>
-         <column>
-          <property name="text">
-           <string>Repository</string>
-          </property>
-         </column>
-         <column>
-          <property name="text">
-           <string>Returncode</string>
-          </property>
-         </column>
-        </widget>
-       </item>
-        <item row="2" column="0">
-    <widget class="QLabel" name="logLink">
-        <property name="text">
-            <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;&lt;a href=&quot;file:///&quot;&gt;&lt;span style=&quot; text-decoration: underline; color:#0984e3;&quot;&gt;View the logs&lt;/span&gt;&lt;/a&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
-        </property>
-        <property name="indent">
-            <number>0</number>
-        </property>
-        <property name="openExternalLinks">
-            <bool>true</bool>
-        </property>
-    </widget>
-</item>
+      <layout class="QGridLayout" name="logTableLayout">
+
+
       </layout>
-     </widget>
+      </widget>
      <widget class="QWidget" name="page_3">
       <property name="geometry">
        <rect>
