+++ src/vorta/views/schedule_tab.py
@@ -1,30 +1,17 @@
 from PyQt6 import QtCore, uic
 from PyQt6.QtCore import QDateTime, QLocale, Qt
-from PyQt6.QtWidgets import (
-    QAbstractItemView,
-    QApplication,
-    QHeaderView,
-    QListWidgetItem,
-    QTableWidgetItem,
-)
+from PyQt6.QtWidgets import QApplication, QListWidgetItem, QVBoxLayout, QWidget
 
-from vorta import application, config
+from vorta import application
 from vorta.i18n import get_locale
 from vorta.scheduler import ScheduleStatusType
-from vorta.store.models import BackupProfileMixin, EventLogModel, WifiSettingModel
+from vorta.store.models import BackupProfileMixin, WifiSettingModel
 from vorta.utils import get_asset, get_sorted_wifis
+from vorta.views.logtablewidget import LogTableWidget
 from vorta.views.utils import get_colored_icon
 
 uifile = get_asset('UI/scheduletab.ui')
 ScheduleUI, ScheduleBase = uic.loadUiType(uifile)
-
-
-class LogTableColumn:
-    Time = 0
-    Category = 1
-    Subcommand = 2
-    Repository = 3
-    ReturnCode = 4
 
 
 class ScheduleTab(ScheduleBase, ScheduleUI, BackupProfileMixin):
@@ -40,40 +27,15 @@ class ScheduleTab(ScheduleBase, ScheduleUI, BackupProfileMixin):
             'fixed': self.scheduleFixedRadio,
         }
 
-        # Set up log table
-        self.logTableWidget.setAlternatingRowColors(True)
-        header = self.logTableWidget.horizontalHeader()
-        self.logLink.setText(
-            f'<a href="file://{config.LOG_DIR}"><span style="text-decoration:'
-            'underline; color:#0984e3;">Click here</span></a> for complete logs.'
-        )
-        header.setVisible(True)
-        [header.setSectionResizeMode(i, QHeaderView.ResizeMode.ResizeToContents) for i in range(5)]
-        header.setSectionResizeMode(3, QHeaderView.ResizeMode.Stretch)
-        self.logTableWidget.setSelectionBehavior(QAbstractItemView.SelectionBehavior.SelectRows)
-        self.logTableWidget.setEditTriggers(QAbstractItemView.EditTrigger.NoEditTriggers)
+        self.logTableWidget = LogTableWidget(self)
 
-        # Scheduler intervals we know
-        self.scheduleIntervalUnit.addItem(self.tr('Minutes'), 'minutes')
-        self.scheduleIntervalUnit.addItem(self.tr('Hours'), 'hours')
-        self.scheduleIntervalUnit.addItem(self.tr('Days'), 'days')
-        self.scheduleIntervalUnit.addItem(self.tr('Weeks'), 'weeks')
+        self.logTableLayout.addWidget(self.logTableWidget)
+        self.logTableWidget.show()
 
-        # Enable/Disable entries on button state changed
-        self.framePeriodic.setEnabled(False)
-        self.frameDaily.setEnabled(False)
-        self.frameValidation.setEnabled(False)
-
-        self.scheduleIntervalRadio.toggled.connect(self.framePeriodic.setEnabled)
-        self.scheduleFixedRadio.toggled.connect(self.frameDaily.setEnabled)
-        self.validationCheckBox.toggled.connect(self.frameValidation.setEnabled)
-
-        # POPULATE with data
         self.populate_from_profile()
         self.set_icons()
 
-        # Connect events
-        self.app.backup_finished_event.connect(self.populate_logs)
+        self.app.backup_finished_event.connect(self.logTableWidget.populate_logs)
 
         # Scheduler events
         for label, obj in self.schedulerRadioMapping.items():
@@ -172,7 +134,7 @@ class ScheduleTab(ScheduleBase, ScheduleUI, BackupProfileMixin):
             self.createCmdLineEdit.setEnabled(False)
 
         self.populate_wifi()
-        self.populate_logs()
+        self.logTableWidget.populate_logs()
         self.draw_next_scheduled_backup()
 
     def draw_next_scheduled_backup(self):
@@ -218,22 +180,3 @@ class ScheduleTab(ScheduleBase, ScheduleUI, BackupProfileMixin):
         repo = self.profile().repo
         setattr(repo, attr, new_value)
         repo.save()
-
-    def populate_logs(self):
-        event_logs = [s for s in EventLogModel.select().order_by(EventLogModel.start_time.desc())]
-
-        sorting = self.logTableWidget.isSortingEnabled()
-        self.logTableWidget.setSortingEnabled(False)  # disable sorting while modifying the table.
-        self.logTableWidget.setRowCount(len(event_logs))  # go ahead and set table length and then update the rows
-        for row, log_line in enumerate(event_logs):
-            formatted_time = log_line.start_time.strftime('%Y-%m-%d %H:%M')
-            self.logTableWidget.setItem(row, LogTableColumn.Time, QTableWidgetItem(formatted_time))
-            self.logTableWidget.setItem(row, LogTableColumn.Category, QTableWidgetItem(log_line.category))
-            self.logTableWidget.setItem(row, LogTableColumn.Subcommand, QTableWidgetItem(log_line.subcommand))
-            self.logTableWidget.setItem(row, LogTableColumn.Repository, QTableWidgetItem(log_line.repo_url))
-            self.logTableWidget.setItem(
-                row,
-                LogTableColumn.ReturnCode,
-                QTableWidgetItem(str(log_line.returncode)),
-            )
-        self.logTableWidget.setSortingEnabled(sorting)  # restore sorting now that modifications are done
